* extrasrc.h, misc: added the missing URL_Open() inline stub so that the   MorphOS compile proceeds again. Also fixed some minor compiler warnings.

+++ src/Threads.c
@@ -160,8 +160,8 @@ static BOOL InitThreadTimer(struct Thread *thread)
 
   if((thread->timer_port = AllocSysObjectTags(ASOT_PORT, TAG_DONE)) != NULL)
   {
-    if((thread->timer_req = (struct TimeRequest *)AllocSysObjectTags(ASOT_IOREQUEST, ASOIOR_Size, sizeof(struct TimeRequest),
-                                                                                     ASOIOR_ReplyPort, thread->timer_port,
+    if((thread->timer_req = (struct TimeRequest *)AllocSysObjectTags(ASOT_IOREQUEST, ASOIOR_Size,      sizeof(struct TimeRequest),
+                                                                                     ASOIOR_ReplyPort, (ULONG)thread->timer_port,
                                                                                      TAG_DONE)) != NULL)
     {
       if(OpenDevice(TIMERNAME, UNIT_VBLANK, (struct IORequest *)thread->timer_req, 0) == 0)
